% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkContingencyMatrix.R
\name{checkContingencyMatrix}
\alias{checkContingencyMatrix}
\title{Yield issue test for contingency matrix.}
\usage{
checkContingencyMatrix(matrix)
}
\arguments{
\item{matrix}{a contingency matrix as described in createContingencyMatrix()}
}
\value{
true if the tests are passed, false otherwise
}
\description{
Checks if a contingency matrix is significant for a yield issue.
The test is passed if no columns sums to zero - that is both the
equipment under test and the others processed at least one wafer
and if the yield of the equipment under test is lower than the
average.
}

