% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiTestEquipment.R
\name{chiTestEquipment}
\alias{chiTestEquipment}
\title{Chi square test on contingency matrix}
\usage{
chiTestEquipment(contingencyMatrix)
}
\arguments{
\item{contingencyMatrix}{a contingency matrix as described in createContingencyMatrix()}
}
\value{
a data.frame as in the description
}
\description{
Perform a chi square test on a contingency matrix and saves the result in
a data.frame with 6 columns: 
VALUE is the chi sqare value
BAD_EQUIPMENT is the number of bad wafers processed by the equipment the contingency matrix refers to
GOOD_EQUIPMENT is the number of bad wafers processed by the equipment the contingency matrix refers to
BAD_OTHER is the number of bad wafers processed by other equipments
GOOD_OTHER is the number of good wafers processed by other equipments
P_VALUE is the p value of the chi square statistic
}

