% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createContingencyMatrix.R
\name{createContingencyMatrix}
\alias{createContingencyMatrix}
\title{Create contingency matrix from wafer data}
\usage{
createContingencyMatrix(thisEqWafers, otherEqWafers, badWafers)
}
\arguments{
\item{thisEqWafers}{a data.frame containing at least two columns: LOT, 
that is the lot of the wafer, and QUANTITY, that is the number of wafer in that lot.
It is supposed to contain all the wafers processed by a single equipment.}

\item{otherEqWafers}{a data.frame containing at least two columns: LOT, 
that is the lot of the wafer, and QUANTITY, that is the number of wafer in that lot.
It is supposed to contain all the wafers processed by other equipment.}

\item{badWafers}{a data.frame containing at least two columns: LOT, 
that is the lot of the wafer, and QUANTITY, that is the number of wafer in that lot.
It is supposed to contain all the wafers to be considered bad.}
}
\value{
the contingency matrix
}
\description{
Create a contingency matrix m where:
m[1, 1] = bad wafers processed by an equipment
m[1, 2] = good wafers processed by an equipment
m[2, 1] = bad wafers processed by other equipments
m[2, 2] = good wafers processed by other equipments
}

